Nits, modern-web-monad is close to replacing nginx for my uses.

+++ src/private-event-server.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <unordered_map>
+
 #include "openssl/ssl.h"
 #include "openssl/err.h"
 
@@ -12,7 +14,7 @@ private:
 	bool recv(int fd, char* data, size_t data_length);
 	bool nonblocking_accept();
 public:
-	PrivateEventServer(std::string certificate, std::string private_key, uint16_t port, int max_connections);
+	PrivateEventServer(std::string certificate, std::string private_key, uint16_t port, size_t max_connections);
 	~PrivateEventServer();
 	bool send(int fd, const char* data, int data_length);
 };
